Refactor track selection handling and introduce PlaybackTracksState for improved media track management
